+++ opps/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from __future__ import unicode_literals
 import pkg_resources
 
 
@@ -7,12 +8,33 @@ pkg_resources.declare_namespace(__name__)
 VERSION = (0, 3, 0)
 __version__ = ".".join(map(str, VERSION))
 __status__ = "Development"
-__description__ = u"Open Source Content Management Platform - CMS for the "
-u"magazines, newspappers websites and portals with "
-u"high-traffic, using the Django Framework."
+__description__ = "Open Source Content Management Platform - CMS for the "
+"magazines, newspappers websites and portals with "
+"high-traffic, using the Django Framework."
 
-__author__ = u"Thiago Avelino"
+__author__ = "Thiago Avelino"
 __credits__ = ['Bruno Rocha']
-__email__ = u"thiago@avelino.xxx"
-__license__ = u"MIT"
-__copyright__ = u"Copyright 2014, Thiago Avelino"
+__email__ = "thiago@avelino.xxx"
+__license__ = "MIT"
+__copyright__ = "Copyright 2014, Thiago Avelino"
+
+
+OPPS_CORE_APPS = [
+    # Django core
+    'django.contrib.admin',
+    'django.contrib.auth',
+    'django.contrib.contenttypes',
+    'django.contrib.sessions',
+    'django.contrib.messages',
+    'django.contrib.staticfiles',
+    'django.contrib.sites',
+    'django.contrib.redirects',
+
+    # Dependence
+    'treebeard',
+
+    # Opps Core
+    'opps.core',
+    'opps.channels',
+    'opps.containers',
+]